Entire application now sits behind authentication 🎉

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
   async createUser(parent, args, {
@@ -24,28 +25,63 @@ const Mutation = {
       }, 'thisisasecret')
     }
   },
-  async deleteUser(parent, args, {
+  async login(parent, args, {
     prisma
   }, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email
+      }
+    })
+
+    if (!user) {
+      throw new Error('Unable to login')
+    }
+
+    const isMatch = await bcrypt.compare(args.data.password, user.password)
+
+    if (!isMatch) {
+      throw new Error('Unable to login')
+    }
+
+    return {
+      user,
+      token: jwt.sign({
+        userId: user.id
+      }, 'thisisasecret')
+    }
+  },
+  async deleteUser(parent, args, {
+    prisma,
+    request
+  }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.deleteUser({
       where: {
-        id: args.id
+        id: userId
       }
     }, info)
   },
   async updateUser(parent, args, {
-    prisma
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
     }, info)
   },
   createPost(parent, args, {
-    prisma
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+
     return prisma.mutation.createPost({
       data: {
         title: args.data.title,
@@ -53,24 +89,64 @@ const Mutation = {
         published: args.data.published,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         }
       }
     }, info)
   },
-  deletePost(parent, args, {
-    prisma
+  async deletePost(parent, args, {
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+
     return prisma.mutation.deletePost({
       where: {
         id: args.id
       }
     }, info)
   },
-  updatePost(parent, args, {
-    prisma
+  async updatePost(parent, args, {
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    const isPublished = await prisma.exists.Post({
+      id: args.id,
+      published: true
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
+
+    if (isPublished && args.data.published === false) {
+      await prisma.mutation.deleteManyComments({
+        where: {
+          post: {
+            id: args.id
+          }
+        }
+      })
+    }
+
     return prisma.mutation.updatePost({
       where: {
         id: args.id
@@ -78,15 +154,26 @@ const Mutation = {
       data: args.data
     }, info)
   },
-  createComment(parent, args, {
-    prisma
+  async createComment(parent, args, {
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.data.post,
+      published: true
+    })
+
+    if (!postExists) {
+      throw new Error('Unable to find post')
+    }
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         },
         post: {
@@ -97,18 +184,44 @@ const Mutation = {
       }
     }, info)
   },
-  deleteComment(parent, args, {
-    prisma
+  async deleteComment(parent, args, {
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to delete comment')
+    }
+
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
       }
     }, info)
   },
-  updateComment(parent, args, {
-    prisma
+  async updateComment(parent, args, {
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) {
+      throw new Error('Unable to update comment')
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id: args.id
